rework sso after auth0 action, pre-user-registration
